requirements.txt
backoff==1.8.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-bigquery==1.16.0
google-cloud-core==1.0.2
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
jsonschema==2.6.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2018.4
requests==2.22.0
rsa==4.0
simplejson==3.11.1
singer-python==5.7.0
six==1.12.0
-e git+https://github.com/anelendata/tap_bigquery.git@113560179525164710bcc9d617c915e1a8e3c978#egg=tap_bigquery
urllib3==1.25.3
